lesson28/content-moderation-system/ml-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
scikit-learn==1.4.2
transformers==4.40.1
torch==2.3.0
numpy==1.26.4
pandas==2.2.2
redis==5.0.4
requests==2.31.0
python-multipart==0.0.9
pillow==10.3.0
opencv-python==4.9.0.80
tensorflow==2.16.1
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.1
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
